requirements.txt
fastapi>=0.95.1
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-xdist>=3.0.0
pytest-cov>=4.0.0
ruff>=0.1.6
pre-commit>=3.0.0
openai~=1.78.1
requests~=2.32.3
psycopg2-binary>=2.9.0
pgvector>=0.2.5
